Move classes into src folder
