+++ src/main/java/com/ford/henrysgroceries/products/ProductHelper.java
@@ -1,4 +1,4 @@
-package com.ford.products;
+package com.ford.henrysgroceries.products;
 
 import java.math.BigDecimal;
